# src/main/java/com/studentmanagement/springrest/api/StudentdetailsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (4.3.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.studentmanagement.springrest.api;

import com.studentmanagement.springrest.api.model.StudentDetails;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.CookieValue;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-05-24T11:49:50.730915900+05:30[Asia/Calcutta]")

@Validated
@Api(value = "studentdetails", description = "the studentdetails API")
public interface StudentdetailsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /studentdetails : Getting all student details
     *
     * @return A list of all student details (status code 200)
     *         or Successful (status code 201)
     *         or Bad request input (status code 400)
     *         or Data not found (status code 404)
     *         or Invalid input (status code 500)
     *         or Bad Gatway (status code 502)
     */
    @ApiOperation(value = "Getting all student details", nickname = "studentdetailsGet", notes = "", response = StudentDetails.class, tags={ "Student Information", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "A list of all student details", response = StudentDetails.class),
        @ApiResponse(code = 201, message = "Successful"),
        @ApiResponse(code = 400, message = "Bad request input"),
        @ApiResponse(code = 404, message = "Data not found"),
        @ApiResponse(code = 500, message = "Invalid input"),
        @ApiResponse(code = 502, message = "Bad Gatway") })
    @RequestMapping(value = "/studentdetails",
        produces = { "applicantion/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<StudentDetails> _studentdetailsGet() {
        return studentdetailsGet();
    }

    // Override this method
    default  ResponseEntity<StudentDetails> studentdetailsGet() {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf(""))) {
                    String exampleString = "";
                    ApiUtil.setExampleResponse(request, "", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /studentdetails/{id} : Getting student details by id
     *
     * @param id  (required)
     * @return A list of a student details (status code 200)
     *         or Successful (status code 201)
     *         or Bad request input (status code 400)
     *         or Data not found (status code 404)
     *         or Invalid input (status code 500)
     *         or Bad Gatway (status code 502)
     */
    @ApiOperation(value = "Getting student details by id", nickname = "studentdetailsIdGet", notes = "", response = StudentDetails.class, tags={ "Student Information", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "A list of a student details", response = StudentDetails.class),
        @ApiResponse(code = 201, message = "Successful"),
        @ApiResponse(code = 400, message = "Bad request input"),
        @ApiResponse(code = 404, message = "Data not found"),
        @ApiResponse(code = 500, message = "Invalid input"),
        @ApiResponse(code = 502, message = "Bad Gatway") })
    @RequestMapping(value = "/studentdetails/{id}",
        produces = { "applicantion/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<StudentDetails> _studentdetailsIdGet(@ApiParam(value = "",required=true) @PathVariable("id") Integer id) {
        return studentdetailsIdGet(id);
    }

    // Override this method
    default  ResponseEntity<StudentDetails> studentdetailsIdGet(Integer id) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf(""))) {
                    String exampleString = "";
                    ApiUtil.setExampleResponse(request, "", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /studentdetails/{id} : Editing student details
     *
     * @param id  (required)
     * @param body student details that need to be added (required)
     * @return A list of a student details (status code 200)
     *         or Successful (status code 201)
     *         or Bad request input (status code 400)
     *         or Data not found (status code 404)
     *         or Invalid input (status code 500)
     *         or Bad Gatway (status code 502)
     */
    @ApiOperation(value = "Editing student details", nickname = "studentdetailsIdPut", notes = "", response = StudentDetails.class, tags={ "Student Information", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "A list of a student details", response = StudentDetails.class),
        @ApiResponse(code = 201, message = "Successful"),
        @ApiResponse(code = 400, message = "Bad request input"),
        @ApiResponse(code = 404, message = "Data not found"),
        @ApiResponse(code = 500, message = "Invalid input"),
        @ApiResponse(code = 502, message = "Bad Gatway") })
    @RequestMapping(value = "/studentdetails/{id}",
        produces = { "applicantion/json" }, 
        consumes = { "applicantion/json" },
        method = RequestMethod.PUT)
    default ResponseEntity<StudentDetails> _studentdetailsIdPut(@ApiParam(value = "",required=true) @PathVariable("id") Integer id,@ApiParam(value = "student details that need to be added" ,required=true )  @Valid @RequestBody StudentDetails body) {
        return studentdetailsIdPut(id, body);
    }

    // Override this method
    default  ResponseEntity<StudentDetails> studentdetailsIdPut(Integer id, StudentDetails body) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf(""))) {
                    String exampleString = "";
                    ApiUtil.setExampleResponse(request, "", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /studentdetails : Adding a new student
     *
     * @param body student details that needs to be added (required)
     * @return Successful (status code 201)
     *         or Bad request input (status code 400)
     *         or Data not found (status code 404)
     *         or Invalid input (status code 500)
     *         or Bad Gatway (status code 502)
     */
    @ApiOperation(value = "Adding a new student", nickname = "studentdetailsPost", notes = "", tags={ "Student Information", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "Successful"),
        @ApiResponse(code = 400, message = "Bad request input"),
        @ApiResponse(code = 404, message = "Data not found"),
        @ApiResponse(code = 500, message = "Invalid input"),
        @ApiResponse(code = 502, message = "Bad Gatway") })
    @RequestMapping(value = "/studentdetails",
        consumes = { "application/json" },
        method = RequestMethod.POST)
    default ResponseEntity<Void> _studentdetailsPost(@ApiParam(value = "student details that needs to be added" ,required=true )  @Valid @RequestBody StudentDetails body) {
        return studentdetailsPost(body);
    }

    // Override this method
    default  ResponseEntity<Void> studentdetailsPost(StudentDetails body) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}